+++ magefile.go
@@ -30,13 +30,18 @@ var (
 	// GoLicenserImportPath controls the import path used to install go-licenser.
 	GoLicenserImportPath = "github.com/elastic/go-licenser"
 
-	publicDir    = "./public"
-	buildDir     = "./build"
-	packagePaths = []string{"./dev/packages/alpha/", "./dev/packages/beats/", "./dev/packages/example/"}
-	tarGz        = true
+	publicDir      = "./public"
+	buildDir       = "./build"
+	storageRepoDir = filepath.Join(buildDir, "package-storage")
+	packagePaths   = []string{filepath.Join(storageRepoDir, "packages"), "./dev/packages/example/"}
+	tarGz          = true
 )
 
 func Build() error {
+	packagePathsEnv := os.Getenv("PACKAGE_PATHS")
+	if packagePathsEnv != "" {
+		packagePaths = strings.Split(packagePathsEnv, ",")
+	}
 
 	err := os.RemoveAll(publicDir)
 	if err != nil {
@@ -44,6 +49,11 @@ func Build() error {
 	}
 
 	err = os.MkdirAll(publicDir, 0755)
+	if err != nil {
+		return err
+	}
+
+	err = fetchPackageStorage()
 	if err != nil {
 		return err
 	}
@@ -68,23 +78,36 @@ func Build() error {
 	return sh.Run("go", "build", ".")
 }
 
-func ImportBeats() error {
-	args := []string{"run", "./dev/import-beats/"}
-	if os.Getenv("SKIP_KIBANA") == "true" {
-		args = append(args, "-skipKibana")
+func fetchPackageStorage() error {
+	err := os.RemoveAll(storageRepoDir)
+	if err != nil {
+		return err
 	}
-	args = append(args, "*.go")
-	return sh.Run("go", args...)
+
+	err = sh.Run("git", "clone", "https://github.com/elastic/package-storage.git", storageRepoDir)
+	if err != nil {
+		return err
+	}
+
+	packageStorageRevision := os.Getenv("PACKAGE_STORAGE_REVISION")
+	if packageStorageRevision == "" {
+		packageStorageRevision = "master"
+	}
+
+	return sh.Run("git",
+		"--git-dir", filepath.Join(storageRepoDir, ".git"),
+		"--work-tree", storageRepoDir,
+		"checkout",
+		packageStorageRevision)
 }
 
 // Creates the `index.json` file
 // For now only containing the version.
 func BuildRootFile() error {
 	rootData := map[string]string{
-		"version":      "0.3.0",
+		"version":      "0.4.0",
 		"service.name": "package-registry",
 	}
-
 	return writeJsonFile(rootData, publicDir+"/index.json")
 }
 
@@ -226,8 +249,7 @@ func Clean() error {
 	if err != nil {
 		return err
 	}
-
-	return os.Remove("package-registry")
+	return os.RemoveAll("package-registry")
 }
 
 func Vendor() error {